extract and optimize collectionStats
