+++ src/window_main/createSelect.tsx
@@ -10,7 +10,7 @@ export default function createSelect(
   callback: (option: string) => void,
   divClass: string,
   optionFormatter?: (option: string) => string | JSX.Element
-): Element {
+): HTMLElement {
   // We create this container outside the component to ensure wherever we're rendering doens't get blanked.
   const selectContainer = createDiv(["select_container", divClass]);
   selectContainer.id = divClass;
